basic cache for search page meta data
The data is queried with no parameters, so we we can cache the whole
value and spare us some time.

We can't use the cache from standard library because the functions that
we want to wrap use the hash of the parameter (in our case the session
object, which comes from a pool).

update tests with get_all_datasets mock

clear session cache in integration test

added clear cache in tests

skip acceptance-tests